+++ test/unit/mysql_perf/unit_test_run.sh
@@ -6,7 +6,6 @@
 
 echo ""
 echo "Unit testing..."
-/usr/bin/python test/unit/mysql_perf/convert_dict.py
 /usr/bin/python test/unit/mysql_perf/create_data_config.py
 /usr/bin/python test/unit/mysql_perf/create_header.py
 /usr/bin/python test/unit/mysql_perf/data_out.py
@@ -14,5 +13,4 @@ echo "Unit testing..."
 /usr/bin/python test/unit/mysql_perf/main.py
 /usr/bin/python test/unit/mysql_perf/mysql_stat.py
 /usr/bin/python test/unit/mysql_perf/mysql_stat_run.py
-/usr/bin/python test/unit/mysql_perf/process_json.py
 /usr/bin/python test/unit/mysql_perf/run_program.py