+++ test/unit/sonarqube_code_coverage.sh
@@ -8,7 +8,6 @@ coverage erase
 
 echo ""
 echo "Running unit test modules in conjunction with coverage"
-coverage run -a --source=mysql_perf test/unit/mysql_perf/convert_dict.py
 coverage run -a --source=mysql_perf test/unit/mysql_perf/create_data_config.py
 coverage run -a --source=mysql_perf test/unit/mysql_perf/create_header.py
 coverage run -a --source=mysql_perf test/unit/mysql_perf/data_out.py
@@ -16,7 +15,6 @@ coverage run -a --source=mysql_perf test/unit/mysql_perf/help_message.py
 coverage run -a --source=mysql_perf test/unit/mysql_perf/main.py
 coverage run -a --source=mysql_perf test/unit/mysql_perf/mysql_stat.py
 coverage run -a --source=mysql_perf test/unit/mysql_perf/mysql_stat_run.py
-coverage run -a --source=mysql_perf test/unit/mysql_perf/process_json.py
 coverage run -a --source=mysql_perf test/unit/mysql_perf/run_program.py